Modularized Dashboard and library detail. Had to change main dashboard page to be libraries. Subject to change.
