+++ UI/Web/src/app/collections/collections.module.ts
@@ -2,9 +2,9 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { CollectionDetailComponent } from './collection-detail/collection-detail.component';
 import { SharedModule } from '../shared/shared.module';
-import { CollectionsRoutingModule } from './collections-routing.module';
 import { CardsModule } from '../cards/cards.module';
 import { AllCollectionsComponent } from './all-collections/all-collections.component';
+import { CollectionsRoutingModule } from './collections-routing.module';
 import { SidenavModule } from '../sidenav/sidenav.module';
 
 
@@ -17,9 +17,11 @@ import { SidenavModule } from '../sidenav/sidenav.module';
   imports: [
     CommonModule,
     SharedModule,
+
     CardsModule,
+    SidenavModule,
+
     CollectionsRoutingModule,
-    SidenavModule
   ],
   exports: [
     AllCollectionsComponent
