+++ UI/Web/src/app/dashboard/dashboard.component.html
@@ -1,3 +1,31 @@
 <app-side-nav-companion-bar>
 </app-side-nav-companion-bar>
-<app-library></app-library>
+<ng-container *ngIf="libraries.length === 0 && !isLoading">
+    <div class="mt-3">
+        <div *ngIf="isAdmin" class="d-flex justify-content-center">
+            <p>There are no libraries setup yet. Configure some in <a routerLink="/admin/dashboard" fragment="libraries">Server settings</a>.</p>
+        </div>
+        <div *ngIf="!isAdmin" class="d-flex justify-content-center">
+            <p>You haven't been granted access to any libraries.</p>
+        </div>
+    </div>
+</ng-container>
+
+<app-carousel-reel [items]="inProgress" title="On Deck" (sectionClick)="handleSectionClick($event)">
+    <ng-template #carouselItem let-item let-position="idx">
+        <app-series-card [data]="item" [libraryId]="item.libraryId" [suppressLibraryLink]="libraryId !== 0" (reload)="reloadInProgress($event)" (dataChanged)="reloadInProgress($event)"></app-series-card>
+    </ng-template>
+</app-carousel-reel>
+
+<app-carousel-reel [items]="recentlyUpdatedSeries" title="Recently Updated Series" (sectionClick)="handleSectionClick($event)">
+    <ng-template #carouselItem let-item let-position="idx">
+        <app-card-item [entity]="item" [title]="item.seriesName" [suppressLibraryLink]="libraryId !== 0" [imageUrl]="imageService.getSeriesCoverImage(item.seriesId)"
+        [supressArchiveWarning]="true" (clicked)="handleRecentlyAddedChapterClick(item)" [count]="item.count"></app-card-item>
+    </ng-template>
+</app-carousel-reel>
+
+<app-carousel-reel [items]="recentlyAddedSeries" title="Newly Added Series" (sectionClick)="handleSectionClick($event)">
+    <ng-template #carouselItem let-item let-position="idx">
+        <app-series-card [data]="item" [libraryId]="item.libraryId" [suppressLibraryLink]="libraryId !== 0" (dataChanged)="loadRecentlyAddedSeries()"></app-series-card>
+    </ng-template>
+</app-carousel-reel>
