+++ UI/Web/src/app/reading-list/reading-list.module.ts
@@ -2,7 +2,7 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { DragableOrderedListComponent } from './dragable-ordered-list/dragable-ordered-list.component';
 import { ReadingListDetailComponent } from './reading-list-detail/reading-list-detail.component';
-import { ReadingListRoutingModule } from './reading-list.router.module';
+import { ReadingListRoutingModule } from './reading-list-routing.module';
 import {DragDropModule} from '@angular/cdk/drag-drop';
 import { AddToListModalComponent } from './_modals/add-to-list-modal/add-to-list-modal.component';
 import { ReactiveFormsModule } from '@angular/forms';
@@ -11,8 +11,8 @@ import { ReadingListsComponent } from './reading-lists/reading-lists.component';
 import { EditReadingListModalComponent } from './_modals/edit-reading-list-modal/edit-reading-list-modal.component';
 import { PipeModule } from '../pipe/pipe.module';
 import { SharedModule } from '../shared/shared.module';
-import { SidenavModule } from '../sidenav/sidenav.module';
 import { NgbNavModule } from '@ng-bootstrap/ng-bootstrap';
+import { SidenavModule } from '../sidenav/sidenav.module';
 
 
 
@@ -26,14 +26,17 @@ import { NgbNavModule } from '@ng-bootstrap/ng-bootstrap';
   ],
   imports: [
     CommonModule,
-    ReadingListRoutingModule,
     ReactiveFormsModule,
     DragDropModule,
-    CardsModule,
+    NgbNavModule,
+
     PipeModule,
     SharedModule,
+
+    CardsModule,
     SidenavModule,
-    NgbNavModule
+
+    ReadingListRoutingModule,
   ],
   exports: [
     AddToListModalComponent,
