+++ UI/Web/src/app/sidenav/side-nav/side-nav.component.html
@@ -7,7 +7,7 @@
             </ng-container>
         </app-side-nav-item> -->
 
-        <app-side-nav-item icon="fa-home" title="Home" link="/library/"></app-side-nav-item>
+        <app-side-nav-item icon="fa-home" title="Home" link="/libraries/"></app-side-nav-item>
         <app-side-nav-item icon="fa-list" title="Collections" link="/collections/"></app-side-nav-item>
         <app-side-nav-item icon="fa-list-ol" title="Reading Lists" link="/lists/"></app-side-nav-item>
         <app-side-nav-item icon="fa-bookmark" title="Bookmarks" link="/bookmarks/"></app-side-nav-item>
